+++ packages/atomic/src/components.d.ts
@@ -868,6 +868,7 @@ export namespace Components {
     interface AtomicQuickviewModal {
         "content"?: string;
         "current"?: number;
+        "modalCloseCallback"?: () => void;
         "reset": () => Promise<void>;
         "result"?: Result;
         "sandbox"?: string;
@@ -3516,6 +3517,7 @@ declare namespace LocalJSX {
     interface AtomicQuickviewModal {
         "content"?: string;
         "current"?: number;
+        "modalCloseCallback"?: () => void;
         "onAtomic/quickview/next"?: (event: AtomicQuickviewModalCustomEvent<any>) => void;
         "onAtomic/quickview/previous"?: (event: AtomicQuickviewModalCustomEvent<any>) => void;
         "result"?: Result;