+++ packages/atomic-react/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.26](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@2.1.25...@coveo/atomic-react@2.1.26) (2023-02-10)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+## [2.1.25](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@2.1.24...@coveo/atomic-react@2.1.25) (2023-02-09)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
 ## [2.1.24](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@2.1.23...@coveo/atomic-react@2.1.24) (2023-02-08)
 
 **Note:** Version bump only for package @coveo/atomic-react